+++ src/test/java/nextstep/member/acceptance/AuthAcceptanceSteps.java
@@ -1,6 +1,6 @@
-package nextstep.subway.acceptance;
+package nextstep.member.acceptance;
 
-import static nextstep.subway.acceptance.MemberSteps.베어러_인증_로그인_요청;
+import static nextstep.member.acceptance.MemberSteps.베어러_인증_로그인_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.restassured.response.ExtractableResponse;
