+++ src/test/java/nextstep/member/acceptance/MemberAcceptanceTest.java
@@ -1,16 +1,17 @@
-package nextstep.subway.acceptance;
+package nextstep.member.acceptance;
 
-import static nextstep.subway.acceptance.MemberSteps.베어러_인증_로그인_요청;
-import static nextstep.subway.acceptance.MemberSteps.베어러_인증으로_내_회원_정보_조회_요청;
-import static nextstep.subway.acceptance.MemberSteps.회원_삭제_요청;
-import static nextstep.subway.acceptance.MemberSteps.회원_생성_요청;
-import static nextstep.subway.acceptance.MemberSteps.회원_정보_수정_요청;
-import static nextstep.subway.acceptance.MemberSteps.회원_정보_조회_요청;
-import static nextstep.subway.acceptance.MemberSteps.회원_정보_조회됨;
+import static nextstep.member.acceptance.MemberSteps.베어러_인증_로그인_요청;
+import static nextstep.member.acceptance.MemberSteps.베어러_인증으로_내_회원_정보_조회_요청;
+import static nextstep.member.acceptance.MemberSteps.회원_삭제_요청;
+import static nextstep.member.acceptance.MemberSteps.회원_생성_요청;
+import static nextstep.member.acceptance.MemberSteps.회원_정보_수정_요청;
+import static nextstep.member.acceptance.MemberSteps.회원_정보_조회_요청;
+import static nextstep.member.acceptance.MemberSteps.회원_정보_조회됨;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import nextstep.utils.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
