+++ src/test/java/nextstep/utils/DatabaseCleanup.java
@@ -1,4 +1,4 @@
-package nextstep.subway.utils;
+package nextstep.utils;
 
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.context.annotation.Profile;
